pyproject.toml
[project]
name = "clinical-drug-discovery"
version = "0.1.0"
description = "Clinical-enriched drug discovery using PrimeKG and clinical notes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dagster>=1.11.16",
    "dagster-postgres>=0.27.16",
    "dagster-webserver>=1.11.16",
    "mlflow>=3.5.1",
    "neo4j>=6.0.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "scispacy>=0.6.2",
    "spacy>=3.8.7",
    "tqdm>=4.67.1",
]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
